fix(client): read_from client API was incorrectly using provided length value as an end index
- Minor refactoring in sn_api moving the SafeData struct into its own file.
- Re-enabling the only two fetch API tests which are now passsing with this fix.
